removed unused exception classes, mvn commands, readme file and gitignore
