--- patterns/hexagonal-architecture/src/main/java/com/baeldung/hexagonal/exception/RestExceptionHandler.java
@@ -1,21 +1,0 @@
-package com.baeldung.hexagonal.exception;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-@ControllerAdvice
-public class RestExceptionHandler extends ResponseEntityExceptionHandler {
-
-    @ExceptionHandler({ ApiRequestException.class })
-    public ResponseEntity<Object> handleResponseStatusException(ApiRequestException ex, WebRequest request) {
-        HttpHeaders httpHeaders = new HttpHeaders();
-        httpHeaders.add(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_UTF8_VALUE);
-        return new ResponseEntity<Object>(ex.getApiError(), httpHeaders, ex.getStatus());
-    }
-
-}
