+++ patterns/hexagonal-architecture/src/main/java/com/baeldung/hexagonal/port/BookOutboundPort.java
@@ -1,4 +1,4 @@
-package com.baeldung.hexagonal.port.outbound;
+package com.baeldung.hexagonal.port;
 
 import com.baeldung.hexagonal.domain.Book;
