+++ patterns/hexagonal-architecture/src/test/java/com/baeldung/hexagonal/adapter/BookRepositoryAdapterUnitTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.hexagonal.adapter.outbound;
+package com.baeldung.hexagonal.adapter;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -6,6 +6,7 @@ import static org.mockito.Mockito.timeout;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import com.baeldung.hexagonal.adapter.BookRepositoryAdapter;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
